meshoptimizer
include external reference to meshoptimizer
switch Model::Index to uint32_t this is for meshoptimizer usage.
add tp folder for third party libraries/tools.
add CMake option CSGJS_TEST_MESHOPTIMIZER to enable meshoptimizer test.

+++ csgjs.h
@@ -187,7 +187,7 @@ struct Polygon {
 
 struct Model {
 
-    using Index = uint16_t;
+    using Index = uint32_t;
 
     CSGJSCPP_VECTOR<Vertex> vertices;
     CSGJSCPP_VECTOR<Index>  indices;
